gnss: change TS to uint32_t
The timestamp was now int64_t. This is way too high and we are not
expecting to operate until the end of time. Change it to uint32_t which
still allows for ~50day of continuous operation. By reducing it by 4
bytes, we reduce the LoRa payload by ~20% and decrease our air time to
less than a second

+++ lib/gnss/gnss.cpp
@@ -148,7 +148,8 @@ void gnss_astrolavos_task(void* args)
                 static_cast<uint8_t>(gps.satellites.value()));
             astrolavos_app->updateCoordinates(
                 {static_cast<float>(gps.location.lat()),
-                 static_cast<float>(gps.location.lng()), esp_timer_get_time()});
+                 static_cast<float>(gps.location.lng()),
+                 static_cast<uint32_t>(esp_timer_get_time())});
             gnss_power_down();
             esp_pm_lock_release(lock);
             utils::delay_ms(GNSS_TASK_LOCATED_SLEEP);
